+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Add missing new keyword to `ExcelChart.Title`
+- Dispose crypto provider
 
 ## [1.1.0] - 2021-07-22